front: move local xml processing to xml helper
